\name{getNChains.twDEMC}
\alias{getNChains.twDEMC}
\title{getNChains twDEMC}
\description{Extracts the number of chains}
\usage{
getNChains.twDEMC(res, ...)
}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}

\value{integer, number of populations in twDEMC}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCInt}}}

