\name{logLikDs2MultiTemp}
\alias{logLikDs2MultiTemp}
\title{logLikDs2MultiTemp}
\description{Misfit for theta}
\usage{
logLikDs2MultiTemp(theta, theta0, theta1 = 7, offset = -100, maxy = numeric(0))
}
\arguments{
  \item{theta}{
}
  \item{theta0}{
}
  \item{theta1}{
}
  \item{offset}{
}
  \item{maxy}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





