\name{subset.twDEMC}
\alias{subset.twDEMC}
\title{subset twDEMC}
\description{Condenses an twDEMC List to the cases boKeep.}
\usage{
subset.twDEMC(x, boKeep, ...)
}
\arguments{
  \item{x}{
}
  \item{boKeep}{
}
  \item{\dots}{
}
}

\value{list of class twDEMC with subset of cases}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{subChains.twDEMC}}}

